fix: QA fixes from PR #108/#117 review
- Extract dispatcher utilities from dom-primitives.ts into dom-dispatch.ts
  (reduces dom-primitives.ts by ~220 LOC, keeps script builders self-contained)
- Strip C0 control characters from cleanText in page summary script
- Add TestFormatCommandResult_FinalField with 5 subtests covering the
  final field contract (complete/error → true, pending → false, expired/timeout → isError)
- Update golden file for schema changes
