+++ extension/background/dom-dispatch.js
@@ -1,0 +1,213 @@
+/**
+ * Purpose: Handles extension background coordination and message routing.
+ * Docs: docs/features/feature/analyze-tool/index.md
+ * Docs: docs/features/feature/interact-explore/index.md
+ * Docs: docs/features/feature/observe/index.md
+ */
+import { domPrimitive, domWaitFor, domFrameProbe } from './dom-primitives.js';
+function parseDOMParams(query) {
+    try {
+        return typeof query.params === 'string' ? JSON.parse(query.params) : query.params;
+    }
+    catch {
+        return null;
+    }
+}
+function isReadOnlyAction(action) {
+    return action === 'list_interactive' || action.startsWith('get_');
+}
+function normalizeFrameTarget(frame) {
+    if (frame === undefined || frame === null)
+        return undefined;
+    if (typeof frame === 'number') {
+        if (!Number.isInteger(frame) || frame < 0)
+            return null;
+        return frame;
+    }
+    if (typeof frame === 'string') {
+        const trimmed = frame.trim();
+        if (trimmed.length === 0)
+            return null;
+        return trimmed;
+    }
+    return null;
+}
+async function resolveExecutionTarget(tabId, frame) {
+    const normalized = normalizeFrameTarget(frame);
+    if (normalized === null) {
+        throw new Error('invalid_frame');
+    }
+    if (normalized === undefined || normalized === 'all') {
+        return { tabId, allFrames: true };
+    }
+    const probeResults = await chrome.scripting.executeScript({
+        target: { tabId, allFrames: true },
+        world: 'MAIN',
+        func: domFrameProbe,
+        args: [normalized]
+    });
+    const frameIds = Array.from(new Set(probeResults
+        .filter((r) => !!r.result?.matches)
+        .map((r) => r.frameId)
+        .filter((id) => typeof id === 'number')));
+    if (frameIds.length === 0) {
+        throw new Error('frame_not_found');
+    }
+    return { tabId, frameIds };
+}
+/** Pick the best result from multi-frame executeScript. Prefers main frame, falls back to first success. */
+function pickFrameResult(results) {
+    const mainFrame = results.find((r) => r.frameId === 0);
+    if (mainFrame?.result && mainFrame.result.success) {
+        return { result: mainFrame.result, frameId: 0 };
+    }
+    for (const r of results) {
+        if (r.result && r.result.success) {
+            return { result: r.result, frameId: r.frameId };
+        }
+    }
+    if (mainFrame?.result)
+        return { result: mainFrame.result, frameId: 0 };
+    return results[0] ? { result: results[0].result, frameId: results[0].frameId } : null;
+}
+/** Merge list_interactive results from all frames (up to 100 elements). */
+function mergeListInteractive(results) {
+    const elements = [];
+    for (const r of results) {
+        const res = r.result;
+        if (res?.elements)
+            elements.push(...res.elements);
+        if (elements.length >= 100)
+            break;
+    }
+    return { success: true, elements: elements.slice(0, 100) };
+}
+async function executeWaitFor(target, params) {
+    const selector = params.selector || '';
+    const quickCheck = await chrome.scripting.executeScript({
+        target,
+        world: 'MAIN',
+        func: domPrimitive,
+        args: [params.action, selector, { timeout_ms: params.timeout_ms }]
+    });
+    const quickPicked = pickFrameResult(quickCheck);
+    const quickResult = quickPicked?.result;
+    if (quickResult?.success)
+        return quickResult;
+    return chrome.scripting.executeScript({
+        target,
+        world: 'MAIN',
+        func: domWaitFor,
+        args: [selector, params.timeout_ms || 5000]
+    });
+}
+async function executeStandardAction(target, params) {
+    return chrome.scripting.executeScript({
+        target,
+        world: 'MAIN',
+        func: domPrimitive,
+        args: [
+            params.action,
+            params.selector || '',
+            {
+                text: params.text,
+                value: params.value,
+                clear: params.clear,
+                checked: params.checked,
+                name: params.name,
+                timeout_ms: params.timeout_ms,
+                analyze: params.analyze
+            }
+        ]
+    });
+}
+function sendToastForResult(tabId, readOnly, result, actionToast, toastLabel, toastDetail) {
+    if (readOnly)
+        return;
+    if (result.success) {
+        actionToast(tabId, toastLabel, toastDetail, 'success');
+    }
+    else {
+        actionToast(tabId, toastLabel, result.error || 'failed', 'error');
+    }
+}
+// #lizard forgives
+export async function executeDOMAction(query, tabId, syncClient, sendAsyncResult, actionToast) {
+    const params = parseDOMParams(query);
+    if (!params) {
+        sendAsyncResult(syncClient, query.id, query.correlation_id, 'error', null, 'invalid_params');
+        return;
+    }
+    const { action, selector, reason } = params;
+    if (!action) {
+        sendAsyncResult(syncClient, query.id, query.correlation_id, 'error', null, 'missing_action');
+        return;
+    }
+    if (action === 'wait_for' && !selector) {
+        sendAsyncResult(syncClient, query.id, query.correlation_id, 'error', null, 'missing_selector');
+        return;
+    }
+    const toastLabel = reason || action;
+    const toastDetail = reason ? undefined : selector || 'page';
+    const readOnly = isReadOnlyAction(action);
+    // Enrich successful results with effective tab context (post-execution URL).
+    // Agents compare resolved_url (dispatch time) vs effective_url (execution time) to detect drift.
+    const enrichWithEffectiveContext = async (result) => {
+        try {
+            const tab = await chrome.tabs.get(tabId);
+            if (result && typeof result === 'object' && !Array.isArray(result)) {
+                return { ...result, effective_tab_id: tabId, effective_url: tab.url };
+            }
+            return result;
+        }
+        catch {
+            return result;
+        }
+    };
+    try {
+        const executionTarget = await resolveExecutionTarget(tabId, params.frame);
+        const tryingShownAt = Date.now();
+        if (!readOnly)
+            actionToast(tabId, toastLabel, toastDetail, 'trying', 10000);
+        const rawResult = action === 'wait_for'
+            ? await executeWaitFor(executionTarget, params)
+            : await executeStandardAction(executionTarget, params);
+        // wait_for quick-check can return a DOMResult directly
+        if (!Array.isArray(rawResult)) {
+            if (!readOnly)
+                actionToast(tabId, toastLabel, toastDetail, 'success');
+            sendAsyncResult(syncClient, query.id, query.correlation_id, 'complete', await enrichWithEffectiveContext(rawResult));
+            return;
+        }
+        // Ensure "trying" toast is visible for at least 500ms
+        const MIN_TOAST_MS = 500;
+        const elapsed = Date.now() - tryingShownAt;
+        if (!readOnly && elapsed < MIN_TOAST_MS)
+            await new Promise((r) => setTimeout(r, MIN_TOAST_MS - elapsed));
+        // list_interactive: merge elements from all frames
+        if (action === 'list_interactive') {
+            const merged = mergeListInteractive(rawResult);
+            sendAsyncResult(syncClient, query.id, query.correlation_id, 'complete', await enrichWithEffectiveContext(merged));
+            return;
+        }
+        const picked = pickFrameResult(rawResult);
+        const firstResult = picked?.result;
+        if (firstResult && typeof firstResult === 'object') {
+            const resultPayload = params.frame !== undefined && params.frame !== null && picked
+                ? { ...firstResult, frame_id: picked.frameId }
+                : firstResult;
+            sendToastForResult(tabId, readOnly, resultPayload, actionToast, toastLabel, toastDetail);
+            sendAsyncResult(syncClient, query.id, query.correlation_id, 'complete', await enrichWithEffectiveContext(resultPayload));
+        }
+        else {
+            if (!readOnly)
+                actionToast(tabId, toastLabel, 'no result', 'error');
+            sendAsyncResult(syncClient, query.id, query.correlation_id, 'error', null, 'no_result');
+        }
+    }
+    catch (err) {
+        actionToast(tabId, action, err.message, 'error');
+        sendAsyncResult(syncClient, query.id, query.correlation_id, 'error', null, err.message);
+    }
+}
+//# sourceMappingURL=dom-dispatch.js.map
