+++ src/background/pending-queries.ts
@@ -24,7 +24,7 @@ import * as eventListeners from './event-listeners'
 import * as index from './index'
 import { DebugCategory } from './debug'
 import { saveStateSnapshot, loadStateSnapshot, listStateSnapshots, deleteStateSnapshot } from './message-handlers'
-import { executeDOMAction } from './dom-primitives'
+import { executeDOMAction } from './dom-dispatch'
 import { executeUpload } from './upload-handler'
 import { canTakeScreenshot, recordScreenshot } from './state-manager'
 import { startRecording, stopRecording } from './recording'
